corrosion_detection/corrosion_detection/requirements.txt
folium==0.14.0
pandas
PyMySQL==1.1.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel
tensorflow-io-gcs-filesystem==0.31.0
numpy==1.24.3
langchain==0.0.322
langchainplus-sdk==0.0.20
Flask-Cors
openpyxl
openai==0.28.1
pillow
langchain_experimental
langchain-community
reportlab